apps/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic-settings==2.4.0
python-multipart==0.0.9
httpx==0.27.0
sqlmodel==0.0.22
pillow >= 10.4.0
pgpy==0.6.0
opencv-python-headless >= 4
pytest >= 8
reedsolo >= 1.5.0
PGPy >= 0.5.0
SQLAlchemy >= 2.0.0
psycopg2-binary >= 2.8
web3>=6.11.1
eth-account>=0.10.0
python-dotenv>=1.0.1
pycryptodome
snscrape >= 0.7.0.20230622
requests